Move sorted to erigon-lib
